verdictai-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
xgboost==2.0.3
shap==0.45.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
sentence-transformers==2.7.0
sqlmodel==0.0.16
python-dotenv==1.0.1
httpx==0.27.0
google-auth==2.49.2
aiosmtplib==4.0.2
google-generativeai
groq
python-multipart
